x11 terminal is now less zealous about replotting
Some window managers (notion for instance) produce several ConfigureNotify
events when a window is created. Some of these were mis-interpreted by gnuplot
as a window manager refusing gnuplot's suggested resize, which resulted in
extra, unneeded replot actions. This patch waits 0.1s for extra ConfigureNotify
events before deciding that a resize was rejected. If a new ConfigureNotify is
seen, we save a replot at the cost of 0.1s. If not, we proceed as before.