gplt_x11.c: not trying to display() incomplete images
Previously it was possible for gnuplot to try to call display() before all
chunks in an image have been received. This caused confusion resulting in
corrupted plots. This patch blocks display() calls if we're in the middle of an
image transfer. This issue was discussed on 2013/04/07 on the gnuplot-beta
mailing list:

http://comments.gmane.org/gmane.comp.graphics.gnuplot.devel/11280